// apps/api/src/common/security/access-denied-reason.enum.ts
export enum AccessDeniedReason {
  TEAM_VIEW_FORBIDDEN = 'team_view_forbidden',
  TEAM_MANAGE_FORBIDDEN = 'team_manage_forbidden',
  FILE_PROJECT_ACCESS_FORBIDDEN = 'file_project_access_forbidden',
  FILE_DELETE_NOT_OWNER = 'file_delete_not_owner',
  FILE_OWNER_MISSING = 'file_owner_missing',
  FILE_BULK_DELETE_FORBIDDEN = 'file_bulk_delete_forbidden',
  PRODUCT_PROJECT_ACCESS_FORBIDDEN = 'product_project_access_forbidden',
  PROJECT_UPDATE_NOT_OWNER = 'project_update_not_owner',
  PROJECT_DELETE_NOT_OWNER = 'project_delete_not_owner',
  PROJECT_OWNER_MISSING = 'project_owner_missing',
  PROJECT_VIEW_FORBIDDEN = 'project_view_forbidden',
  PRODUCT_DELETE_NOT_OWNER = 'product_delete_not_owner',
  PRODUCT_OWNER_MISSING = 'product_owner_missing',
  EVENT_VIEW_PRIVATE_FORBIDDEN = 'event_view_private_forbidden',
  EVENT_UPDATE_NOT_OWNER = 'event_update_not_owner',
  EVENT_MANAGE_NOT_OWNER = 'event_manage_not_owner',
  EVENT_DELETE_NOT_OWNER = 'event_delete_not_owner',
  EVENT_OWNER_MISSING = 'event_owner_missing',
  ACTIVITY_PROJECT_ACCESS_FORBIDDEN = 'activity_project_access_forbidden',
  ACTIVITY_DELETE_PROJECT_NOT_OWNER = 'activity_delete_project_not_owner',
  ACTIVITY_DELETE_EVENT_NOT_OWNER = 'activity_delete_event_not_owner',
  ACTIVITY_DELETE_NOT_OWNER = 'activity_delete_not_owner',
  ACTIVITY_OWNER_MISSING = 'activity_owner_missing',
  STANDALONE_PRODUCT_VIEW_FORBIDDEN = 'standalone_product_view_forbidden',
  STANDALONE_PRODUCT_MANAGE_FORBIDDEN = 'standalone_product_manage_forbidden',
}